[wip] Finish FFI separation with other public functions
