+++ rust/tw_memory/src/ffi/c_byte_array.rs
@@ -1,0 +1,25 @@
+// TODO declare `free_c_byte_array`, `is_c_byte_array_null` FFI.
+#[repr(C)]
+pub struct CByteArray {
+    pub data: *mut u8,
+    pub size: usize,
+}
+
+impl From<Vec<u8>> for CByteArray {
+    fn from(value: Vec<u8>) -> Self {
+        let size = value.len();
+        let mut mut_vec = value.clone();
+        let ptr = mut_vec.as_mut_ptr();
+        std::mem::forget(mut_vec);
+        CByteArray { data: ptr, size }
+    }
+}
+
+impl CByteArray {
+    pub fn null() -> CByteArray {
+        CByteArray {
+            data: std::ptr::null_mut(),
+            size: 0,
+        }
+    }
+}
